Break out runtime lib to main and worker
