+++ cli/js.rs
@@ -16,7 +16,8 @@ pub static COMPILER_SNAPSHOT_MAP: &[u8] =
 pub static COMPILER_SNAPSHOT_DTS: &[u8] =
   include_bytes!(concat!(env!("OUT_DIR"), "/COMPILER_SNAPSHOT.d.ts"));
 
-pub static DENO_RUNTIME: &str = include_str!("js/lib.deno_runtime.d.ts");
+pub static DENO_NS_LIB: &str = include_str!("js/lib.deno.d.ts");
+pub static DENO_MAIN_LIB: &str = include_str!("js/lib.deno_main.d.ts");
 
 #[test]
 fn cli_snapshot() {
