+++ cli/js/compiler_bootstrap.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
-import { ASSETS, Host } from "./compiler_host.ts";
+import { ASSETS, CompilerHostTarget, Host } from "./compiler_host.ts";
 import { core } from "./core.ts";
 import * as dispatch from "./dispatch.ts";
 import { getAsset } from "./compiler_util.ts";
@@ -15,8 +15,27 @@ for (const [name, opId] of Object.entries(ops)) {
   (dispatch as any)[opName] = opId;
 }
 
-const host = new Host({ writeFile(): void {} });
+const host = new Host({
+  target: CompilerHostTarget.Main,
+  writeFile(): void {}
+});
 const options = host.getCompilationSettings();
+
+// This is a hacky way of adding our libs to the libs available in TypeScript()
+// as these are internal APIs of TypeScript which maintain valid libs
+/* eslint-disable @typescript-eslint/no-explicit-any */
+(ts as any).libs.push("deno_main", "deno_worker", "deno");
+(ts as any).libMap.set("deno_main", "lib.deno_main.d.ts");
+(ts as any).libMap.set("deno_worker", "lib.deno_worker.d.ts");
+(ts as any).libMap.set("deno", "lib.deno.d.ts");
+/* eslint-enable @typescript-eslint/no-explicit-any */
+
+// this pre-populates the cache at snapshot time of our library files, so they
+// are available in the future when needed.
+host.getSourceFile(`${ASSETS}/lib.deno_main.d.ts`, ts.ScriptTarget.ESNext);
+host.getSourceFile(`${ASSETS}/lib.deno_worker.d.ts`, ts.ScriptTarget.ESNext);
+host.getSourceFile(`${ASSETS}/lib.deno.d.ts`, ts.ScriptTarget.ESNext);
+host.getSourceFile(`${ASSETS}/lib.webworker.d.ts`, ts.ScriptTarget.ESNext);
 
 /** Used to generate the foundational AST for all other compilations, so it can
  * be cached as part of the snapshot and available to speed up startup */
