+++ cli/js/lib.deno_worker.d.ts
@@ -1,0 +1,5 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+/// <reference no-default-lib="true" />
+/// <reference lib="deno" />
+/// <reference lib="webworker" />
